apps/realtime-analyzer/pyproject.toml
[tool.poetry]
name = "realtime-analyzer"
version = "0.1.0"
description = "Realtime EEG analysis service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Webフレームワークと非同期実行
flask = "^3.0.3"
# RabbitMQクライアント
pika = "^1.3.2"
# 脳波解析ライブラリ
mne = "^1.6.1"
mne-connectivity = "^0.6.0" # コヒーレンス計算用
# 数値計算
numpy = "^1.26.4"
# 圧縮・解凍
zstandard = "^0.22.0"
# グラフ描画 (バックエンド指定)
matplotlib = "^3.9.0"
# 環境変数管理
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"